refactor: share connector network transport fields
