+++ frontend/src/connectors/templates/postgres/form.jsx
@@ -1,11 +1,9 @@
 import { Field, Input, Select } from "../../../components/ui/form";
 import { Notice } from "../../../components/ui/notice";
-import { HostPingButton } from "../host-ping-button";
+import { NetworkTransportFields } from "../_shared/network-transport-fields";
 
 export function PostgresConnectorFormTemplate({ form, mode = "create", targets = [], onChange }) {
   const editing = mode === "edit";
-  const sshProfiles = sshProfileOptions(targets);
-  const overSSH = form.connection_mode === "over_ssh";
   return (
     <>
       <Notice tone="good">
@@ -15,50 +13,13 @@ export function PostgresConnectorFormTemplate({ form, mode = "create", targets =
         Connector name
         <Input value={form.name} onChange={(event) => onChange("name", event.target.value)} required />
       </Field>
-      <Field>
-        Connection mode
-        <Select value={form.connection_mode} onChange={(event) => onChange("connection_mode", event.target.value)}>
-          <option value="direct">Direct from this gateway</option>
-          <option value="over_ssh">Over an SSH connector profile</option>
-        </Select>
-      </Field>
-      {overSSH ? (
-        <Field>
-          SSH transport profile
-          <Select value={form.transport_target_ref} onChange={(event) => onChange("transport_target_ref", event.target.value)} required>
-            <option value="" disabled>
-              Select SSH profile
-            </option>
-            {sshProfiles.map((profile) => (
-              <option value={profile.ref} key={profile.ref}>
-                {profile.label}
-              </option>
-            ))}
-          </Select>
-        </Field>
-      ) : null}
-      {overSSH ? (
-        <Notice>
-          Host and port are resolved from the SSH server. Use 127.0.0.1:5432 when Postgres only listens on the remote machine.
-        </Notice>
-      ) : (
-        <Notice>
-          For Postgres running on the same Linux host as AIPermission Docker, use host.docker.internal instead of localhost.
-        </Notice>
-      )}
-      <div className="grid gap-3 sm:grid-cols-[minmax(0,1fr)_120px]">
-        <Field>
-          <span className="flex items-center justify-between gap-2">
-            <span>Host</span>
-            <HostPingButton host={form.host} port={form.port} mode={form.connection_mode} transportTargetRef={form.transport_target_ref} />
-          </span>
-          <Input value={form.host} onChange={(event) => onChange("host", event.target.value)} required />
-        </Field>
-        <Field>
-          Port
-          <Input type="number" min="1" max="65535" value={form.port} onChange={(event) => onChange("port", event.target.value)} required />
-        </Field>
-      </div>
+      <NetworkTransportFields
+        form={form}
+        targets={targets}
+        onChange={onChange}
+        overSSHNotice="Host and port are resolved from the SSH server. Use 127.0.0.1:5432 when Postgres only listens on the remote machine."
+        directNotice="For Postgres running on the same Linux host as AIPermission Docker, use host.docker.internal instead of localhost."
+      />
       <div className="grid gap-3 sm:grid-cols-[minmax(0,1fr)_160px]">
         <Field>
           Database
@@ -101,15 +62,4 @@ export function PostgresConnectorFormTemplate({ form, mode = "create", targets =
       </Field>
     </>
   );
-}
-
-function sshProfileOptions(targets) {
-  return (targets || [])
-    .filter((target) => target.connector_kind === "ssh")
-    .flatMap((target) =>
-      (target.profiles || []).map((profile) => ({
-        ref: profile.ref || `${target.connector_kind}:${target.id}:${profile.id}`,
-        label: `${target.name} / ${profile.label} · ${target.config?.host || "host"}:${target.config?.port || 22}`,
-      }))
-    );
 }
